requirements.txt
Pillow>=10.0.0
pystray>=0.19.0
customtkinter>=5.2.0
darkdetect>=0.8.0
opencv-python>=4.8.0
numpy>=1.24.0
tkinterdnd2>=0.3.0
